On invalid submit, focus the correct tab
On a failed submit, it is entirely possible that the invalid field is not
on the tab that the user is currently viewing. So, in order to not just
fail silently and lead the user to believe that nothing is happening, we
first need to show the tab of the invalid field, and then focus that
field.